+++ src/index.js
@@ -3,7 +3,6 @@ var app = require('./app');
 var samples = require('./samples');
 var cordova = require('./cordova');
 var test = require('./test');
-var publish = require('./publish');
 var api = require('./api');
 var submodules = require('./submodules');
 
@@ -11,7 +10,6 @@ function all() {
     build();
     app();
     samples();
-    publish();
     cordova();
     test();
     api();
@@ -22,7 +20,6 @@ module.exports = all;
 module.exports.build = build,
 module.exports.app = app,
 module.exports.samples = samples;
-module.exports.publish = publish;
 module.exports.cordova = cordova;
 module.exports.test = test;
 module.exports.api = api;