+++ src/publish.js
@@ -1,5 +1,0 @@
-var conf = require('./config');
-
-module.exports = function() {
-    
-};